Extract `CFFParser` from fonts.js (issue 6777)
